+++ src/main/kotlin/net/ossrs/ideasrs/SrsBundle.kt
@@ -8,9 +8,9 @@ import org.jetbrains.annotations.NonNls
 import org.jetbrains.annotations.PropertyKey
 
 @NonNls
-private const val BUNDLE = "messages.MyBundle"
+private const val BUNDLE = "messages.SrsBundle"
 
-object MyBundle : AbstractBundle(BUNDLE) {
+object SrsBundle : AbstractBundle(BUNDLE) {
 
     @Suppress("SpreadOperator")
     @JvmStatic